Reorganize handlers and handler instantiation
