+++ lib/serviceworker/engine.rb
@@ -3,13 +3,14 @@
 require "rails"
 require "rails/railtie"
 require "serviceworker"
+require "serviceworker/handlers/sprockets_handler"
 
 module ServiceWorker
   class Engine < ::Rails::Engine
     config.serviceworker = ActiveSupport::OrderedOptions.new
     config.serviceworker.headers = {}
     config.serviceworker.routes = ServiceWorker::Router.new
-    config.serviceworker.handler = ServiceWorker::Rails::Handler.new
+    config.serviceworker.handler = ServiceWorker::Handlers::SprocketsHandler.new
     config.serviceworker.icon_sizes = %w[36 48 60 72 76 96 120 152 180 192 512]
 
     initializer "serviceworker-rails.configure_rails_initialization", after: :load_config_initializers do
