+++ lib/serviceworker/rails.rb
@@ -4,5 +4,4 @@ module ServiceWorker
 end
 
 require "serviceworker/rails/version"
-require "serviceworker/rails/handler"
 require "serviceworker/engine"
