+++ lib/serviceworker/route.rb
@@ -17,7 +17,7 @@ module ServiceWorker
       end
 
       @path_pattern = path_pattern
-      @asset_pattern = if options[:pack] && defined?(Webpacker)
+      @asset_pattern = if options[:pack] && defined?(::Webpacker)
                          asset_pattern || options.fetch(:pack, path_pattern)
                        else
                          asset_pattern || options.fetch(:asset, path_pattern)
